Extract IDCStudyRowComponent.
